requirements_orig.txt
# Foundational Libraries (core dependencies)
setuptools==59.5.0
numpy==1.23.5            # Pin numpy to ensure compatibility
scipy==1.9.3             # Install scipy after numpy
pandas==1.5.3            # Pin pandas to avoid dask incompatibilities

# Libraries dependent on numpy, pandas, or scipy
numba==0.58.1
dask==2023.4.1           # Use a stable version compatible with pandas==1.5.3
scikit-learn==1.2.2
imageio==2.34.0
matplotlib==3.7.1
seaborn==0.12.2
wordcloud==1.9.3

# Machine Learning and Deep Learning
torchmetrics==0.10.3
pytorch-lightning==1.9.5
transformers==4.29.2
deepspeed==0.9.2
accelerate==0.19.0
diffusers==0.16.1
optuna==3.1.1
fair-esm==2.0.0

# Specialized Libraries
lightning-flash==0.8.2
axial-positional-embedding==0.2.1
linear-attention-transformer==0.19.1
linformer==0.2.3

# Bioinformatics
biopython==1.79

# General Utilities
tqdm==4.65.0
pynvml==11.5.0

# Additions (added by A Howe)
h5py
